Move duplicate code to shared location
